Move each element to separate file
